NMS-12259: Use kinda proper package name
