+++ features/flows/classification/shell/pom.xml
@@ -22,7 +22,7 @@
                         <Bundle-SymbolicName>${project.groupId}.${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Version>${project.version}</Bundle-Version>
                         <Private-Package></Private-Package>
-                        <Karaf-Commands>org.opennms.shell</Karaf-Commands>
+                        <Karaf-Commands>org.opennms.netmgt.flows.clazzification.shell</Karaf-Commands>
                     </instructions>
                 </configuration>
             </plugin>
