+++ features/flows/classification/shell/src/main/java/org/opennms/netmgt/flows/clazzification/shell/ClassificationReloadCommand.java
@@ -26,28 +26,25 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.shell;
+package org.opennms.netmgt.flows.clazzification.shell;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
+import org.apache.karaf.shell.api.action.Action;
+import org.apache.karaf.shell.api.action.Command;
+import org.apache.karaf.shell.api.action.lifecycle.Reference;
 import org.apache.karaf.shell.api.action.lifecycle.Service;
-import org.apache.karaf.shell.api.console.CommandLine;
-import org.apache.karaf.shell.api.console.Completer;
-import org.apache.karaf.shell.api.console.Session;
-import org.apache.karaf.shell.support.completers.StringsCompleter;
-import org.opennms.netmgt.flows.classification.persistence.api.Protocols;
+import org.opennms.netmgt.flows.classification.ClassificationEngine;
 
+@Command(scope="classification", name="reload-engine", description = "Reloads the rules of the classification engine")
 @Service
-public class ProtocolCompleter implements Completer {
+public class ClassificationReloadCommand implements Action {
+
+    @Reference
+    private ClassificationEngine classificationEngine;
 
     @Override
-    public int complete(Session session, CommandLine commandLine, List<String> list) {
-        final StringsCompleter delegate = new StringsCompleter();
-        final List<String> protocols = Protocols.getProtocols().stream()
-                .map(p -> p.getKeyword()).filter(p -> !p.equals("") && !p.equals("Reserved"))
-                .collect(Collectors.toList());
-        delegate.getStrings().addAll(protocols);
-        return delegate.complete(session, commandLine, list);
+    public Object execute() throws Exception {
+        classificationEngine.reload();
+        System.out.println("Classification Engine has been reloaded");
+        return null;
     }
 }
